+++ sitepulse-engine-http-api/src/main/java/com/sitepulse/engine/http/metrics/dto/WeeklyMetricView.java
@@ -1,5 +1,6 @@
 package com.sitepulse.engine.http.metrics.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -16,8 +17,8 @@ import lombok.ToString;
 public class WeeklyMetricView {
 
     private String weekStart;
-    private double progressDelta;
-    private double activityIndex;
+    private BigDecimal progressDelta;
+    private BigDecimal activityIndex;
     private double activeHours;
     private String riskLevel;
 }
